refactor suspend-user
