+++ src/services/suspend-user.ts
@@ -2,12 +2,19 @@ import renderDelete from '../remote/activitypub/renderer/delete';
 import { renderActivity } from '../remote/activitypub/renderer';
 import { deliver } from '../queue';
 import config from '../config';
+import deleteFollowing from '../services/following/delete';
 import { User } from '../models/entities/user';
-import { Users, Followings } from '../models';
+import { Users, Followings, Notifications } from '../models';
 import { Not, IsNull, Brackets } from 'typeorm';
 import { processStreamingRows } from '../misc/process-streaming-rows';
 
 export async function doPostSuspend(user: User) {
+	await unFollowAll(user).catch(e => {});
+	await readAllNotify(user).catch(e => {});
+	await sendDeleteActivity(user).catch(e => {});
+}
+
+export async function sendDeleteActivity(user: User) {
 	if (Users.isLocalUser(user)) {
 		// 知り得る全SharedInboxにDelete配信
 		const content = renderActivity(renderDelete(`${config.url}/users/${user.id}`, user));
@@ -41,4 +48,31 @@ export async function doPostSuspend(user: User) {
 			}
 		});
 	}
+}
+
+async function unFollowAll(follower: User) {
+	const followings = await Followings.find({
+		followerId: follower.id
+	});
+
+	for (const following of followings) {
+		const followee = await Users.findOne({
+			id: following.followeeId
+		});
+
+		if (followee == null) {
+			throw `Cant find followee ${following.followeeId}`;
+		}
+
+		await deleteFollowing(follower, followee, true);
+	}
+}
+
+async function readAllNotify(notifier: User) {
+	await Notifications.update({
+		notifierId: notifier.id,
+		isRead: false,
+	}, {
+		isRead: true
+	});
 }
